Move domain suggestions logic to DomainPickerButton
