+++ client/landing/gutenboarding/components/domain-picker/button.tsx
@@ -15,6 +15,9 @@ import { STORE_KEY as ONBOARD_STORE } from '../../stores/onboard';
 import { isFilledFormValue } from '../../stores/onboard/types';
 
 const DomainPickerButton: FunctionComponent = () => {
+	// // User can search for a domain
+	const [ domainSearch, setDomainSearch ] = useState( '' );
+
 	const [ isDomainPopoverVisible, setDomainPopoverVisibility ] = useState(
 		true /* @TODO: should be `false` by default, true for dev */
 	);
@@ -22,16 +25,21 @@ const DomainPickerButton: FunctionComponent = () => {
 	// Without user search, we can provide recommendations based on title + vertical
 	const { siteTitle, siteVertical } = useSelect( select => select( ONBOARD_STORE ).getState() );
 
+	let search = domainSearch.trim();
+	if ( ! search && isFilledFormValue( siteTitle ) ) {
+		search = siteTitle;
+	}
+
 	const suggestions = useSelect(
 		select => {
-			if ( isFilledFormValue( siteTitle ) ) {
+			if ( isFilledFormValue( search ) ) {
 				return select( DOMAIN_STORE ).getDomainSuggestions( siteTitle, {
 					include_wordpressdotcom: true,
 					...( isFilledFormValue( siteVertical ) && { vertical: siteVertical.id } ),
 				} );
 			}
 		},
-		[ siteTitle, siteVertical ]
+		[ search, siteVertical ]
 	);
 
 	const domainName = suggestions?.[ 1 ]?.domain_name;
@@ -45,7 +53,11 @@ const DomainPickerButton: FunctionComponent = () => {
 					onClick={ e => e.stopPropagation() }
 					onKeyDown={ e => e.stopPropagation() }
 				>
-					<DomainPicker />
+					<DomainPicker
+						domainSearch={ domainSearch }
+						setDomainSearch={ setDomainSearch }
+						suggestions={ suggestions }
+					/>
 				</Popover>
 			) }
 		</Button>
